SKALE-2273 block rotation support in IMA

+++ proxy/contracts/interfaces/IERC721Module.sol
@@ -5,7 +5,7 @@ interface IERC721Module {
     function receiveERC721(
         address contractHere,
         address to,
-        uint tokenId,
+        uint256 tokenId,
         bool isRaw) external returns (bytes memory);
     function sendERC721(address to, bytes calldata data) external returns (bool);
     function getReceiver(address to, bytes calldata data) external pure returns (address);
