+++ proxy/contracts/predeployed/LockAndDataForSchainERC20.sol
@@ -22,9 +22,9 @@ pragma solidity ^0.5.3;
 import "./PermissionsForSchain.sol";
 
 interface ERC20MintAndBurn {
-    function balanceOf(address to) external view returns (uint);
-    function mint(address to, uint amount) external returns (bool);
-    function burn(uint amount) external;
+    function balanceOf(address to) external view returns (uint256);
+    function mint(address to, uint256 amount) external returns (bool);
+    function burn(uint256 amount) external;
 }
 
 
@@ -33,28 +33,28 @@ contract LockAndDataForSchainERC20 is PermissionsForSchain {
     event SendedERC20(bool result);
     event ReceivedERC20(bool result);
 
-    mapping(uint => address) public erc20Tokens;
-    mapping(address => uint) public erc20Mapper;
+    mapping(uint256 => address) public erc20Tokens;
+    mapping(address => uint256) public erc20Mapper;
 
 
     constructor(address _lockAndDataAddress) PermissionsForSchain(_lockAndDataAddress) public {
         // solium-disable-previous-line no-empty-blocks
     }
 
-    function sendERC20(address contractHere, address to, uint amount) external allow("ERC20Module") returns (bool) {
+    function sendERC20(address contractHere, address to, uint256 amount) external allow("ERC20Module") returns (bool) {
         require(ERC20MintAndBurn(contractHere).mint(to, amount), "Could not mint ERC20 Token");
         emit SendedERC20(true);
         return true;
     }
 
-    function receiveERC20(address contractHere, uint amount) external allow("ERC20Module") returns (bool) {
+    function receiveERC20(address contractHere, uint256 amount) external allow("ERC20Module") returns (bool) {
         require(ERC20MintAndBurn(contractHere).balanceOf(address(this)) >= amount, "Amount not transfered");
         ERC20MintAndBurn(contractHere).burn(amount);
         emit ReceivedERC20(true);
         return true;
     }
 
-    function addERC20Token(address addressERC20, uint contractPosition) external allow("ERC20Module") {
+    function addERC20Token(address addressERC20, uint256 contractPosition) external allow("ERC20Module") {
         erc20Tokens[contractPosition] = addressERC20;
         erc20Mapper[addressERC20] = contractPosition;
     }
