+++ proxy/test/contracts/TestSkaleVerifier.sol
@@ -5,12 +5,12 @@ pragma solidity ^0.5.0;
 contract SkaleVerifier {
 
     function verifySchainSignature(
-        uint signA,
-        uint signB,
+        uint256 signA,
+        uint256 signB,
         bytes32 hash,
-        uint counter,
-        uint hashA,
-        uint hashB,
+        uint256 counter,
+        uint256 hashA,
+        uint256 hashB,
         string calldata schainName
     )
         external
